Provider/Backend framework
