+++ metriq_gym/cli.py
@@ -2,6 +2,54 @@
 
 import argparse
 
+from tabulate import tabulate
+
+from metriq_gym.job_manager import JobManager
+from metriq_gym.providers.provider import ProviderType
+
+
+def list_jobs(args: argparse.Namespace, job_manager: JobManager) -> int:
+    """List jobs recorded in the job manager.
+
+    Args:
+        args: Parsed arguments.
+        job_manager: Job manager instance.
+    Returns:
+        Return code.
+    """
+    # Retrieve all jobs from JobManager.
+    jobs = job_manager.get_jobs()
+
+    # Apply filters if specified.
+    if args.filter and args.value:
+        jobs = [job for job in jobs if str(job.get(args.filter, "")).lower() == args.value.lower()]
+
+    # Display jobs in a tabular format.
+    if not jobs:
+        print("No jobs found.")
+        return 0
+
+    # Prepare data for tabulation.
+    headers = ["ID", "Backend", "Type", "Provider", "Misc"]
+    table = [
+        [
+            job.get("id", ""),
+            job.get("backend", ""),
+            job.get("benchmark_name", ""),
+            job.get("provider", ""),
+            ", ".join(
+                f"{key}: {job[key]}"
+                for key in ["qubits", "shots"]
+                if key in job and job[key] is not None
+            ),
+        ]
+        for job in jobs
+    ]
+
+    # Print the table.
+    print(tabulate(table, headers=headers, tablefmt="grid"))
+    return 0
+
 
 def parse_arguments() -> argparse.Namespace:
     """
@@ -33,7 +81,7 @@ def parse_arguments() -> argparse.Namespace:
         "-p",
         "--provider",
         type=str,
-        choices=["ibmq", "quantinuum", "ionq"],
+        choices=ProviderType.list(),
         default="ibmq",
         help="String identifier for backend provider service",
     )
