+++ metriq_gym/schemas/examples/quantum_volume.example.json
@@ -1,7 +1,7 @@
 {
   "benchmark_name": "Quantum Volume",
-  "num_qubits": 5,
-  "shots": 10,
-  "trials": 10,
+  "num_qubits": 2,
+  "shots": 2,
+  "trials": 2,
   "confidence_level": 0.95
 }
